Time: 1417 ms (24.28%) | Memory: 33.2 MB (90.85%) - LeetSync

+++ 560-subarray-sum-equals-k/subarray-sum-equals-k.cpp
@@ -3,7 +3,7 @@ public:
     int subarraySum(vector<int>& nums, int k)
     {
         int n = nums.size();
-        double ans = 0;
+        int ans = 0;
         for( int  i =0; i < n; i++)
         {
             int sum = nums[i];
